+++ core/module_test.go
@@ -11,10 +11,10 @@ type Game struct {
 	Dst ServiceID
 }
 
-func (g *Game) OnRequestMSG(src ServiceID, rid int, data ...interface{}) {
+func (g *Game) OnRequestMSG(src ServiceID, rid uint64, data ...interface{}) {
 	g.Respond(src, rid, "world")
 }
-func (g *Game) OnCallMSG(src ServiceID, cid int, data ...interface{}) {
+func (g *Game) OnCallMSG(src ServiceID, cid uint64, data ...interface{}) {
 	g.Ret(src, cid, "world")
 }
